Refactor Test Configuration to be Created in Preview Column of Builder
