+++ src/pages/assessment-builder/AssessmentBuilder.module.css
@@ -405,7 +405,57 @@
   color: #555;
 }
 
+.form-group-test-container {
+  display: flex;
+  flex-direction: column;
+  row-gap: 0.8rem;
+}
+
+.form-group-test-preview {
+  display: flex;
+  flex-direction: row;
+  gap: 2.5rem;
+  align-items: flex-start;
+}
+
+.form-group-test-preview-label {
+  width: 100%;
+  font-size: 1rem;
+  font-weight: 600;
+  color: #374151;
+  border-bottom: 1px solid #e5e7eb;
+}
+
+.test-params-fields-col {
+  display: flex;
+  flex: 1;
+  flex-direction: column;
+  gap: 0.2rem;
+  max-width: 48%;
+}
+
+.test-preview-modal-col {
+  display: flex;
+  flex: 1 1 0;
+  flex-direction: column;
+  gap: 0.5rem;
+  max-width: 48%;
+}
+
+@media (width <= 900px) {
+  .form-group-test-preview {
+    flex-direction: column;
+    gap: 1rem;
+  }
+
+  .test-params-fields-col,
+  .test-preview-modal-col {
+    max-width: 100%;
+  }
+}
+
 .form-group-test {
+  width: 49%;
   margin-bottom: 0.1rem;
 }
 
@@ -414,6 +464,36 @@
   font-size: 0.875rem;
   font-weight: 500;
   color: #555;
+}
+
+.form-group-parameter {
+  width: 100%;
+}
+
+.evidence-parameter-option {
+  display: flex;
+  gap: 0.3rem;
+  align-items: center;
+  height: 1.2rem;
+}
+
+.form-group-parameter label,
+.evidence-parameter-option label {
+  margin-bottom: 0.2rem;
+  font-size: 0.875rem;
+  font-weight: 500;
+  color: #555;
+}
+
+.test-preview-modal input {
+  padding: 0.25rem 0.5rem;
+  font-size: 0.875rem;
+  border: 1px solid #d1d5db;
+  border-radius: 0.375rem;
+}
+
+.test-preview-modal textarea {
+  resize: none;
 }
 
 .form-control {
@@ -809,7 +889,7 @@
   }
 }
 
-@media (width < 768px) {
+@media (width < 1060px) {
   .builder-layout {
     flex-direction: column;
   }
@@ -1331,10 +1411,10 @@
 }
 
 .test-methods-list {
-  max-height: 200px;
+  max-height: 380px;
   overflow-y: auto;
   background: #fff;
-  border: 1px solid #e5e7eb;
+  border: 1px solid #dcdee3;
   border-radius: 0.375rem;
 }
 
@@ -1356,16 +1436,16 @@
 }
 
 .test-preview-modal {
-  margin-top: 1rem;
-  margin-bottom: 2rem;
-  box-shadow: 0 2px 4px rgb(0 0 0 / 20%);
+  margin-bottom: 3rem;
+  box-shadow: 1px 2px 4px rgb(0 0 0 / 20%);
 }
 
 .test-method-item {
-  padding: 0.5rem 0.75rem;
+  padding: 0.75rem;
+  font-size: 0.875rem;
   font-weight: 500;
   cursor: pointer;
-  border-bottom: 1px solid #f3f4f6;
+  border-bottom: 1px solid #e1e3e6;
 }
 
 .test-method-item:last-child {
@@ -1373,7 +1453,7 @@
 }
 
 .test-method-item:hover {
-  background-color: #f9fafb;
+  background-color: #eef0f2;
 }
 
 .test-method-item.selected {
@@ -1382,8 +1462,7 @@
 }
 
 .method-name {
-  margin-bottom: 0.25rem;
-  font-size: 0.8rem;
+  font-size: 0.875rem;
   color: #374151;
 }
